feat,refactor(frontend): refactor app structure and code style, add routing and fix docker-compose to install dependencies properly
