+++ docker-compose.yml
@@ -17,7 +17,7 @@ services:
     # container_name: frontend-37
     volumes:
       - ./frontend-37:/app
-      - /app/node_modules
+      # - /app/node_modules
     ports:
       - "5173:5173"
