+++ frontend-37/src/components/ui/Button.jsx
@@ -1,5 +1,6 @@
-import styles from "./Button.module.css"
+import styles from "./Button.module.css";
 
 export function Button(props) {
-    return <button className={styles.button} {...props} />
+  return <button className={styles.button} {...props} />;
 }
+
