+++ frontend-37/src/components/ui/Button.module.css
@@ -1,19 +1,19 @@
-
 .button {
-    width: 218px;
-    height: 69px;
-    text-align: center;
-    border-radius: 21px;
-    background-color: #6968FF;
-    color: #FFFFFF;
-    font-size: 24px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: none;
-    font-weight: 400;
+  width: 218px;
+  height: 69px;
+  text-align: center;
+  border-radius: 21px;
+  background-color: #6968ff;
+  color: #ffffff;
+  font-size: 24px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border: none;
+  font-weight: 400;
 }
 
 .button:hover {
-    cursor: pointer;
+  cursor: pointer;
 }
+
