--- frontend-37/src/components/ui/TopBar.module.css
@@ -1,57 +1,0 @@
-.topbar {
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: space-between;
-    position: sticky;
-
-    background-color: white;
-    border-radius: 22px;
-
-    box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-
-    margin: 15px;
-    padding: 21px 28px 21px 28px;
-}
-
-.topbar a {
-    color: inherit;
-    text-decoration: none;
-}
-
-.topbar p {
-    margin: 0;
-}
-
-.topbar-title {
-    display: flex;
-    justify-self: flex-start;
-    flex-basis: 249px;
-
-    font-size: 24px;
-    font-weight: bold;
-
-    flex-flow: column nowrap;
-    justify-content: center;
-}
-
-.topbar-content {
-    display: flex;
-    justify-self: flex-end;
-    flex-basis: 454.54px;
-
-    flex-flow: row wrap;
-    justify-content: space-between;
-
-    font-size: 24px;
-}
-
-@media (width < 800px) {
-    .topbar {
-        gap: 22px;
-    }
-
-    .topbar-content {
-        justify-content: space-evenly;
-        gap: 22px;
-    }
-}
