--- frontend-37/src/components/ui/TopBarElement.jsx
@@ -1,7 +1,0 @@
-import styles from "./TopBarElement.module.css";
-
-export function TopBarElement({children, ...props}){
-    return <div className={styles.topbarElement} {...props}>
-        {children}
-    </div>;
-}
