--- frontend-37/src/components/ui/TopBarElement.module.css
@@ -1,6 +1,0 @@
-.topbarElement{
-    display: flex;
-    flex-flow: column nowrap;
-    
-    justify-content: center;
-}
