+++ frontend-37/src/index.css
@@ -1,9 +1,12 @@
-@import url('https://fonts.googleapis.com/css2?family=Onest:wght@400;500;600;700&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Onest:wght@400;500;600;700&display=swap");
+
+@import url("https://fonts.googleapis.com/css2?family=Azeret+Mono:ital,wght@0,100..900;1,100..900&display=swap");
 
 * {
-  font-family: 'Onest', sans-serif;
+  font-family: "Onest", sans-serif;
 }
 
 body {
   margin: 0;
 }
+
